Fix relative paths in moved actions and rebuild them
